Reorder and update tests
