+++ packages/back-end/src/services/experimentNotifications.ts
@@ -1,10 +1,10 @@
 import { includeExperimentInPayload, getSnapshotAnalysis } from "shared/util";
+import { getMetricResultStatus } from "shared/experiments";
 import {
-  getHealthSettings,
-  getMetricResultStatus,
   getExperimentResultStatus,
-} from "shared/experiments";
-import { orgHasPremiumFeature } from "shared/enterprise";
+  getHealthSettings,
+  orgHasPremiumFeature,
+} from "shared/enterprise";
 import { StatsEngine } from "back-end/types/stats";
 import { Context } from "back-end/src/models/BaseModel";
 import { createEvent, CreateEventData } from "back-end/src/models/EventModel";
